Cleanup to create FormalStatusSelector
The naming of the formallyOkSelector is rather strange when it can be
used for setting a status to be formally not okay, or for setting it to
undetermined.  A more obvious name would be the FormalStatusSelector,
which can then have either of these states.

We should verify who else uses the shared configuration and rename the
logic there as well.  It seems we're sufficiently deep down the rabbit
hole to ignore that step for now.